my-flask-app/requirements.txt
Flask
Flask-SQLAlchemy
redis
mysql-connector-python
